Move BlockRecordDB to BlockStore.
