+++ chia/full_node/block_store.py
@@ -3,14 +3,18 @@ from __future__ import annotations
 import dataclasses
 import logging
 import sqlite3
+from dataclasses import dataclass
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
 
 import typing_extensions
 import zstd
 
-from chia.consensus.block_record import BlockRecord, BlockRecordDB
+from chia.consensus.block_record import BlockRecord
+from chia.types.blockchain_format.classgroup import ClassgroupElement
+from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.serialized_program import SerializedProgram
 from chia.types.blockchain_format.sized_bytes import bytes32
+from chia.types.blockchain_format.sub_epoch_summary import SubEpochSummary
 from chia.types.full_block import FullBlock
 from chia.types.weight_proof import SubEpochChallengeSegment, SubEpochSegments
 from chia.util.db_wrapper import DBWrapper2, execute_fetchone
@@ -22,10 +26,107 @@ from chia.util.full_block_utils import (
     generator_from_block,
     plot_filter_info_from_block,
 )
-from chia.util.ints import uint32
+from chia.util.ints import uint8, uint32, uint64, uint128
 from chia.util.lru_cache import LRUCache
+from chia.util.streamable import Streamable, streamable
 
 log = logging.getLogger(__name__)
+
+
+@streamable
+@dataclass(frozen=True)
+class BlockRecordDB(Streamable):
+    """
+    This class contains the fields from `BlockRecord` that get stored in the DB.
+    Unlike `BlockRecord`, this should never extend with more fields, in order to avoid DB corruption.
+    """
+
+    header_hash: bytes32
+    prev_hash: bytes32
+    height: uint32
+    weight: uint128
+    total_iters: uint128
+    signage_point_index: uint8
+    challenge_vdf_output: ClassgroupElement
+    infused_challenge_vdf_output: Optional[ClassgroupElement]
+    reward_infusion_new_challenge: bytes32
+    challenge_block_info_hash: bytes32
+    sub_slot_iters: uint64
+    pool_puzzle_hash: bytes32
+    farmer_puzzle_hash: bytes32
+    required_iters: uint64
+    deficit: uint8
+    overflow: bool
+    prev_transaction_block_height: uint32
+    timestamp: Optional[uint64]
+    prev_transaction_block_hash: Optional[bytes32]
+    fees: Optional[uint64]
+    reward_claims_incorporated: Optional[List[Coin]]
+    finished_challenge_slot_hashes: Optional[List[bytes32]]
+    finished_infused_challenge_slot_hashes: Optional[List[bytes32]]
+    finished_reward_slot_hashes: Optional[List[bytes32]]
+    sub_epoch_summary_included: Optional[SubEpochSummary]
+
+    def to_block_record(self, pos_ss_cc_challenge_hash: bytes32, cc_sp_hash: bytes32) -> BlockRecord:
+        return BlockRecord(
+            header_hash=self.header_hash,
+            prev_hash=self.prev_hash,
+            height=self.height,
+            weight=self.weight,
+            total_iters=self.total_iters,
+            signage_point_index=self.signage_point_index,
+            challenge_vdf_output=self.challenge_vdf_output,
+            infused_challenge_vdf_output=self.infused_challenge_vdf_output,
+            reward_infusion_new_challenge=self.reward_infusion_new_challenge,
+            challenge_block_info_hash=self.challenge_block_info_hash,
+            sub_slot_iters=self.sub_slot_iters,
+            pool_puzzle_hash=self.pool_puzzle_hash,
+            farmer_puzzle_hash=self.farmer_puzzle_hash,
+            required_iters=self.required_iters,
+            deficit=self.deficit,
+            overflow=self.overflow,
+            prev_transaction_block_height=self.prev_transaction_block_height,
+            pos_ss_cc_challenge_hash=pos_ss_cc_challenge_hash,
+            cc_sp_hash=cc_sp_hash,
+            timestamp=self.timestamp,
+            prev_transaction_block_hash=self.prev_transaction_block_hash,
+            fees=self.fees,
+            reward_claims_incorporated=self.reward_claims_incorporated,
+            finished_challenge_slot_hashes=self.finished_challenge_slot_hashes,
+            finished_infused_challenge_slot_hashes=self.finished_infused_challenge_slot_hashes,
+            finished_reward_slot_hashes=self.finished_reward_slot_hashes,
+            sub_epoch_summary_included=self.sub_epoch_summary_included,
+        )
+
+    @classmethod
+    def from_block_record(cls, block_record: BlockRecord) -> BlockRecordDB:
+        return cls(
+            header_hash=block_record.header_hash,
+            prev_hash=block_record.prev_hash,
+            height=block_record.height,
+            weight=block_record.weight,
+            total_iters=block_record.total_iters,
+            signage_point_index=block_record.signage_point_index,
+            challenge_vdf_output=block_record.challenge_vdf_output,
+            infused_challenge_vdf_output=block_record.infused_challenge_vdf_output,
+            reward_infusion_new_challenge=block_record.reward_infusion_new_challenge,
+            challenge_block_info_hash=block_record.challenge_block_info_hash,
+            sub_slot_iters=block_record.sub_slot_iters,
+            pool_puzzle_hash=block_record.pool_puzzle_hash,
+            farmer_puzzle_hash=block_record.farmer_puzzle_hash,
+            required_iters=block_record.required_iters,
+            deficit=block_record.deficit,
+            overflow=block_record.overflow,
+            prev_transaction_block_height=block_record.prev_transaction_block_height,
+            timestamp=block_record.timestamp,
+            prev_transaction_block_hash=block_record.prev_transaction_block_hash,
+            fees=block_record.fees,
+            reward_claims_incorporated=block_record.reward_claims_incorporated,
+            finished_challenge_slot_hashes=block_record.finished_challenge_slot_hashes,
+            finished_infused_challenge_slot_hashes=block_record.finished_infused_challenge_slot_hashes,
+            finished_reward_slot_hashes=block_record.finished_reward_slot_hashes,
+            sub_epoch_summary_included=block_record.sub_epoch_summary_included,
+        )
 
 
 @typing_extensions.final
@@ -193,7 +294,7 @@ class BlockStore:
 
     async def add_full_block(self, header_hash: bytes32, block: FullBlock, block_record: BlockRecord) -> None:
         self.block_cache.put(header_hash, block)
-        block_record_db: BlockRecordDB = block_record.to_block_record_db()
+        block_record_db: BlockRecordDB = BlockRecordDB.from_block_record(block_record)
 
         if self.db_wrapper.db_version == 2:
             ses: Optional[bytes] = (
@@ -476,8 +577,7 @@ class BlockStore:
                 raise ValueError(f"Header hash {hh} not in the blockchain")
             plot_filter_info = await self.get_plot_filter_info(hh)
             assert plot_filter_info is not None
-            block_record = BlockRecord.from_block_record_db(
-                all_blocks[hh],
+            block_record = all_blocks[hh].to_block_record(
                 plot_filter_info.pos_ss_cc_challenge_hash,
                 plot_filter_info.cc_sp_hash,
             )
@@ -578,8 +678,7 @@ class BlockStore:
 
         plot_filter_info = await self.get_plot_filter_info(header_hash)
         assert plot_filter_info is not None
-        block_record = BlockRecord.from_block_record_db(
-            block_record_db,
+        block_record = block_record_db.to_block_record(
             plot_filter_info.pos_ss_cc_challenge_hash,
             plot_filter_info.cc_sp_hash,
         )
@@ -619,8 +718,7 @@ class BlockStore:
         for hh, block_record_db in ret.items():
             plot_filter_info = await self.get_plot_filter_info(hh)
             assert plot_filter_info is not None
-            block_record = BlockRecord.from_block_record_db(
-                block_record_db,
+            block_record = block_record_db.to_block_record(
                 plot_filter_info.pos_ss_cc_challenge_hash,
                 plot_filter_info.cc_sp_hash,
             )
@@ -705,8 +803,7 @@ class BlockStore:
         for hh, block_record_db in ret.items():
             plot_filter_info = await self.get_plot_filter_info(hh)
             assert plot_filter_info is not None
-            block_record = BlockRecord.from_block_record_db(
-                block_record_db,
+            block_record = block_record_db.to_block_record(
                 plot_filter_info.pos_ss_cc_challenge_hash,
                 plot_filter_info.cc_sp_hash,
             )
